// src/app/shared/types/moameleh.ts
export interface MoamelehType {
    _id?: string;
    tarikhMoameleh?: string;
    tedadSahmMoameleh: number;
    arzeshSahmMoameleh: number;
    forushandeh_username: string;
    forushandeh_fullName: string;
    forushandeh_darkhastId: string;
    kharidar_username: string;
    kharidar_fullName: string;
    kharidar_darkhastId: string;
    userIdSabtKonandeh?: string;
    usernameSabtKonandeh?: string;
    fullnameSabtKonandeh?: string;
    shenasehMoameleh?: string;
}